pyproject.toml
[project]
name = "oauth-master"
version = "0.1.0"
description = "A RAG tool that indexes the OAuth 2.0 spec and references"
authors = [
    {name = "Carlos Frias",email = "carlos.a.frias@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "llama-index (>=0.12.14)",
    "html2text (>=2024.2.26)",
    "tiktoken (>=0.8.0,<0.9.0)",
    "llama-index-readers-web (>=0.3.5,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "llama-index-llms-openai (>=0.3.15,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

